Remove hacbs specifics from integration tests
